Adding in small build test and incorporating Manjesh's suggestions
