+++ tsdef/LocalDataTrackOptions.d.ts
@@ -1,0 +1,7 @@
+import { LocalTrackOptions } from './LocalTrackOptions';
+
+export interface LocalDataTrackOptions extends LocalTrackOptions {
+  maxPacketLifeTime?: number | null;
+  maxRetransmits?: number | null;
+  ordered?: boolean;
+}
