+++ tsdef/LocalTrackOptions.d.ts
@@ -1,0 +1,9 @@
+import { LogLevel, LogLevels } from './types';
+
+export interface LocalTrackOptions {
+  /**
+   * @deprecated
+   */
+    logLevel: LogLevel | LogLevels;
+    name?: string;
+  }
